Separated UI and Game manager to different handlers
