+++ Assets/Scripts/PauseMenu.cs
@@ -13,7 +13,7 @@ public class PauseMenu : MonoBehaviour
     // Start is called before the first frame update
     void Start()
     {
-        float volume = 0.5f;
+        float volume = 1f;
         if (PlayerPrefs.HasKey("Volume")) volume = PlayerPrefs.GetFloat("Volume");
         backgroundMusic.volume = volume;
         pauseMenu.SetActive(false);
