+++ Assets/Scripts/UiManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6c04984776cc838448fc3c160169882a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
