+++ Assets/Tests/PlayMode/GameManagerPlayTest.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using System.Reflection;
@@ -10,43 +11,26 @@ public class GameManagerPlay
 {
 
     [UnityTest]
-    public IEnumerator RestartScene_When_Player_Health_Is_Zero()
-    {
-        int currentScene = SceneManager.GetActiveScene().buildIndex;
-        var gameManagerObject = new GameObject();
-        var gameManager = gameManagerObject.AddComponent<GameManager>();
-        
-        // Mockoljuk a player-t a GameManager-hez
-        var playerObject = new GameObject();
-        var player = playerObject.AddComponent<Player>();
-        playerObject.AddComponent<Camera>();
-        
-        player.health = 0;
-
-
-        yield return null;
-
-        Assert.AreEqual(currentScene, SceneManager.GetActiveScene().buildIndex); // Példa: Itt mockolhatjuk a jelenetváltást.
-    }
-
-    [UnityTest]
     public IEnumerator OnInvaderKilled_Should_Increase_Score_And_Destroy_Invader()
     {
-        var gameManagerObject = new GameObject();
-        var gameManager = gameManagerObject.AddComponent<GameManager>();
-
-        var invaderObject = new GameObject();
-        var invader = invaderObject.AddComponent<Invader>();
-        invader.health = 1;
-        invader.score = 100;
-
-        gameManager.OnInvaderKilled(invader);
+        // You have to load a scene otherwise play mode test fails
+        SceneManager.LoadScene("test");
 
         yield return null;
 
-        Assert.AreEqual(0, invader.health); // Az invader életereje 0
-        Assert.AreEqual(100, gameManager.score); // A pontszám növekedett
-        Assert.IsTrue(invaderObject == null); // Az invader megsemmisült
+        var gameManagerObj = GameObject.Find("GameManager");
+        var invaderObj = GameObject.Find("Invader_01");        
+
+        var gamemanagerComp = gameManagerObj.GetComponent<GameManager>();
+        var invaderComp = invaderObj.GetComponent<Invader>();
+
+        gamemanagerComp.OnInvaderKilled(invaderComp);
+
+        yield return null;
+
+        Assert.AreEqual(0, invaderComp.health); 
+        Assert.AreEqual(10, gamemanagerComp.score); 
+        Assert.IsTrue(invaderObj == null); 
     }
 
 }
