+++ Assets/UiManager.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 8623cec9f14f6384582913ec6bdea622
-DefaultImporter:
+guid: fb00b62a8d624b94aa588bce332465e0
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
