+++ ProjectSettings/EditorBuildSettings.asset
@@ -35,7 +35,7 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/Mission9-alex.unity
     guid: 4507119d73277f044a1e6dec6fb47f15
-  - enabled: 0
+  - enabled: 1
     path: Assets/Scenes/test.unity
-    guid: 2cda990e2423bbf4892e6590ba056729
+    guid: 5bf9acfb79cfc3d4ca757eae47a34da6
   m_configObjects: {}
